# docs/api/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "AuthMessageResponse": {
        "description": "Standardized response message container.",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "status": {
            "default": "success",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "AuthMessageResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PasswordResetConfirmPayload": {
        "description": "Payload to confirm password reset with raw token and new password.",
        "properties": {
          "new_password": {
            "maxLength": 128,
            "minLength": 8,
            "title": "New Password",
            "type": "string"
          },
          "token": {
            "minLength": 1,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token",
          "new_password"
        ],
        "title": "PasswordResetConfirmPayload",
        "type": "object"
      },
      "PasswordResetRequestPayload": {
        "description": "Payload to initiate a password reset.",
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "PasswordResetRequestPayload",
        "type": "object"
      },
...
      "ResendVerificationRequest": {
        "description": "Payload to request resending email verification link.",
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "ResendVerificationRequest",
        "type": "object"
      },
...
      "TokenResponse": {
        "description": "JWT Token response container.",
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "expires_in"
        ],
        "title": "TokenResponse",
        "type": "object"
      },
...
      "UserLoginRequest": {
        "description": "Payload for account login.",
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "UserLoginRequest",
        "type": "object"
      },
      "UserRegisterRequest": {
        "description": "Payload to register a new user account.",
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "full_name": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "password": {
            "maxLength": 128,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "UserRegisterRequest",
        "type": "object"
      },
      "UserResponse": {
        "description": "Public user identity response payload.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "is_verified": {
            "title": "Is Verified",
            "type": "boolean"
          },
          "last_login_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login At"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "full_name",
          "role",
          "is_active",
          "is_verified",
          "created_at",
          "last_login_at"
        ],
        "title": "UserResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "An Agentic Planetary Risk Intelligence Platform. Phase 1 — foundational API skeleton.",
    "title": "GaiaOS",
    "version": "0.5.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "description": "Authenticate user credentials and issue a JWT access token.\n\nRejects inactive or unverified users. Never logs passwords or secrets.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Invalid credentials, inactive account, or unverified email"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "v1",
          "Auth"
        ]
      }
    },
...
    "/api/v1/auth/register": {
      "post": {
        "description": "Register a new user account.\n\nEnforces password strength policy and duplicate email checks.\nIssues a hashed verification token and triggers verification email delivery.",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "Email already exists"
          },
          "422": {
            "description": "Password policy or validation error"
          }
        },
        "summary": "Register",
        "tags": [
          "v1",
          "Auth"
        ]
      }
    },
    "/api/v1/auth/request-reset": {
      "post": {
        "description": "Initiates password reset flow. Returns HTTP 202 with generic response regardless of whether the email exists to prevent account enumeration.",
        "operationId": "request_password_reset_api_v1_auth_request_reset_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequestPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthMessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request a password reset link",
        "tags": [
          "v1",
          "Auth"
        ]
      }
    },
    "/api/v1/auth/resend-verification": {
      "post": {
        "description": "Resend email verification link for unverified accounts.",
        "operationId": "resend_verification_api_v1_auth_resend_verification_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendVerificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthMessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resend Verification",
        "tags": [
          "v1",
          "Auth"
        ]
      }
    },
    "/api/v1/auth/reset": {
      "post": {
        "description": "Reset account password using plaintext token and new password.\n\nSecurity features:\n1. Validates password strength policy.\n2. Hashes new password with Argon2id.\n3. Single transaction boundary: validates token, updates user password, consumes token,\n   and invalidates ALL remaining active tokens for the user atomically (SELECT FOR UPDATE).",
        "operationId": "reset_password_api_v1_auth_reset_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetConfirmPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthMessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Invalid or expired token"
          },
          "422": {
            "description": "Password policy violation"
          }
        },
        "summary": "Reset password using valid reset token",
        "tags": [
          "v1",
          "Auth"
        ]
      }
    },
    "/api/v1/auth/verify-email": {
      "get": {
        "description": "Verify account email via URL verification token (GET link).",
        "operationId": "verify_email_api_v1_auth_verify_email_get",
        "parameters": [
          {
            "description": "Email verification token",
            "in": "query",
            "name": "token",
            "required": true,
            "schema": {
              "description": "Email verification token",
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthMessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Invalid or expired verification token"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Email",
        "tags": [
          "v1",
          "Auth"
        ]
      }
    },
...
  }
}